refactor: Create updater spi package
