+++ src/main/java/org/opentripplanner/updater/alert/GtfsRealtimeAlertsUpdaterParameters.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.updater.alert;
 
 import java.util.Map;
-import org.opentripplanner.updater.PollingGraphUpdaterParameters;
+import org.opentripplanner.updater.spi.PollingGraphUpdaterParameters;
 
 public record GtfsRealtimeAlertsUpdaterParameters(
   String configRef,
