+++ src/main/java/org/opentripplanner/updater/spi/GenericXmlDataSource.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater;
+package org.opentripplanner.updater.spi;
 
 import java.util.List;
 import java.util.Map;
