+++ src/main/java/org/opentripplanner/updater/trip/PollingTripUpdaterParameters.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.updater.trip;
 
 import java.util.Map;
-import org.opentripplanner.updater.PollingGraphUpdaterParameters;
+import org.opentripplanner.updater.spi.PollingGraphUpdaterParameters;
 
 public record PollingTripUpdaterParameters(
   String configRef,
