+++ src/main/java/org/opentripplanner/updater/trip/metrics/StreamingTripUpdateMetrics.java
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.Objects;
 import org.opentripplanner.model.UpdateError;
 import org.opentripplanner.model.UpdateSuccess;
-import org.opentripplanner.updater.UpdateResult;
+import org.opentripplanner.updater.spi.UpdateResult;
 import org.opentripplanner.updater.trip.UrlUpdaterParameters;
 
 /**
