+++ src/main/java/org/opentripplanner/updater/vehicle_parking/VehicleParkingDataSourceFactory.java
@@ -9,7 +9,7 @@ import org.opentripplanner.ext.vehicleparking.parkapi.CarParkAPIUpdater;
 import org.opentripplanner.ext.vehicleparking.parkapi.ParkAPIUpdaterParameters;
 import org.opentripplanner.model.calendar.openinghours.OpeningHoursCalendarService;
 import org.opentripplanner.routing.vehicle_parking.VehicleParking;
-import org.opentripplanner.updater.DataSource;
+import org.opentripplanner.updater.spi.DataSource;
 
 /**
  * Class that can be used to return a custom vehicle parking {@link DataSource}.
