+++ src/main/java/org/opentripplanner/updater/vehicle_parking/VehicleParkingUpdaterParameters.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.updater.vehicle_parking;
 
-import org.opentripplanner.updater.PollingGraphUpdaterParameters;
+import org.opentripplanner.updater.spi.PollingGraphUpdaterParameters;
 
 /**
  * Class that implements {@link PollingGraphUpdaterParameters} and can be extended to include other
